move mixin interfaces to a package
